requirements.txt
appdirs
asgiref
attrs
autopep8
beautifulsoup4
CacheControl
certifi
chardet
colorama
contextlib2
distlib
distro
dj-database-url
Django
django-cors-headers
django-heroku
djangorestframework
gunicorn
html5lib
idna
ipaddr
lockfile
lxml
msgpack
packaging
pathlib-mate
pep517
progress
psycopg2
pycodestyle
pyparsing
pytoml
pytz
requests
retrying
six
soupsieve
SQLAlchemy
sqlparse
toml
urllib3
uszipcode
webencodings
whitenoise